直播列表调整

+++ app/src/main/java/com/jzh/parents/datamodel/data/CategoryData.kt
@@ -6,7 +6,7 @@ package com.jzh.parents.datamodel.data
  * @author ding
  * Created by Ding on 2018/9/17.
  */
-data class CategoryData(val id: Long = 0,
+data class CategoryData(val id: Int = 0,
                         val name: String? = "",
                         val icon: String? = "",
                         val banner: String? = "",
